+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Changelog
 
 - Default to not having a custom tag for classes
+- Added the ability to drag and drop skills, bab, cmb, defenses, concentration and caster level from an actor's sheet to the macro hotbar to quickly create a macro for it
 
 ## 0.74.0